perf: add policy selector, NVP selector, derive selectedTab directly
