+++ src/libs/actions/IOU/index.ts
@@ -1235,8 +1235,7 @@ function initMoneyRequest({
     const newTransactionID = CONST.IOU.OPTIMISTIC_TRANSACTION_ID;
     const currency = policy?.outputCurrency ?? personalPolicy?.outputCurrency ?? CONST.CURRENCY.USD;
 
-    // Disabling this line since currentDate can be an empty string
-    const created = currentDate || format(new Date(), 'yyyy-MM-dd');
+    const created = currentDate ?? format(new Date(), 'yyyy-MM-dd');
 
     // We remove draft transactions created during multi scanning if there are some
     removeDraftTransactions(true, draftTransactions);
